Add codec specific tests for AudioDecoder. Add mulaw/alaw support.
This expands our AudioDecoder test coverage as was done for the
VideoDecoder. Namely we now have a audioDecoder-codec-specific.js
file which is parameterized based on each supported codec.

This adds testing for AAC (ADTS), AAC (MP4), MP3, MULAW, and
ALAW. Vorbis and FLAC will be added in a followup.

R=sandersd

Fixed: 1203825
Change-Id: I8130ee0d607e2f6968471ee671650b70d1269980